+++ src/Marvin.Migrations.PostgreSQL/Changelog.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [0.2.1] - 2018-11-02
+
+### Added
+
+- Additional methods to check database and table existence
+
 ## [0.2] - 2018-11-02
 
-### Add
+### Added
 
 - Added factory to create `PostgreDbProvider`
 - Added `PostgreDbProviderOptions` to configure provider