+++ src/Marvin.Migrations/Changelog.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [0.2.1] - 2018-11-02
+
+### Added
+
+- Additional methods to check database and table existence
+
 ## [0.2] - 2018-11-02
 
-### Add
+### Added
 
 - Added premigration scripts
 - Added options for `IDbProvider` and factory to create `IDbProvider`